// src/app/api/company/[company-id]/dto/update-company.dto.ts
export interface UpdateCompanyDto {
  name?: string;

  description_ka?: string;

  description_eng?: string;

  address?: string;

  phone?: string;

  email?: string;

  website?: string;

  facebook?: string;

  telegram?: string;

  instagram?: string;

  youtube?: string;

  twitter?: string;

  linkedin?: string;

  profile_picture?: string;

  profile_picture_url?: string;
}